requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==2.0.12
Django==4.0
django-jazzmin==2.4.8
django-phonenumber-field==7.0.2
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
gunicorn==20.1.0
idna==3.4
load-dotenv==0.1.0
phonenumbers==8.13.5
PyJWT==2.7.0
python-dotenv==0.19.2
pytz==2021.3
requests==2.26.0
rest-framework-simplejwt==0.0.2
sqlparse==0.4.4
urllib3==1.26.15
